accounted for refactor
