+++ app/models/card.py
@@ -1,4 +1,4 @@
-from ..extensions import db
+from ..dao import db
 from sqlalchemy import UUID
 
 class Card(db.Model):
