+++ app/models/user.py
@@ -1,4 +1,4 @@
-from ..extensions import db
+from ..dao import db
 from sqlalchemy import UUID
 
 class User(db.Model):
